+++ R/is_wine.R
@@ -1,4 +1,4 @@
-#' Checks whether running R on Windows using Wine or not
+#' Check if running R on Windows using Wine
 #'
 #' @return A logical
 #'